pyproject.toml
[tool.poetry]
name = "imagine"
version = "0.1.0"
description = "example of imagine autogenerated code"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "*"
django-filter = "*"
django-environ = "*"
drf-yasg = "*"
djangorestframework = "*"
mysqlclient = "*"
gunicorn = "*"

[tool.poetry.dev-dependencies]
coverage = "*"
factory-boy = "*"
django-stubs = "*"
mypy = "*"
djangorestframework-stubs = "*"
autopep8 = "*"
flake8 = "*"
isort = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
